// dfp-piao-fe/src/app/shared/models/classes/tabella-funzionale-dto.ts
export class TabellaFunzionaleDTO {
  id?: number;
  idEntitaFK?: number;
  codTipologiaFK?: string;
  codice?: string;
  idOVP?: number;
  denominazioneSintetica?: string;
  responsabileAmministrativo?: string;
  idStakeholder?: number;
  dimensioni?: string;
  formula?: string;
  polarita?: string;
  baseline?: string;
  targetAnnoN1?: string;
  targetAnnoN2?: string;
  targetAnnoN3?: string;
  fonte?: string;
}