// dfp-piao-fe/src/app/shared/models/enums/section.enum.ts
export enum SectionEnum {
  SEZIONE_1 = 'SEZIONE_1',
  SEZIONE_2_1 = 'SEZIONE_21',
  SEZIONE_2_2 = 'SEZIONE_22',
  SEZIONE_2_3 = 'SEZIONE_23',
  SEZIONE_3_1 = 'SEZIONE_31',
  SEZIONE_3_2 = 'SEZIONE_32',
  SEZIONE_3_3_1 = 'SEZIONE_331',
  SEZIONE_3_3_2 = 'SEZIONE_332',
  SEZIONE_4 = 'SEZIONE_4',
  PIAO = 'PIAO',
  SEZIONE_21_INDICATORE = 'SEZIONE_21_INDICATORE',
  SEZIONE_22_ADEMPIMENTO = 'SEZIONE_22_ADEMPIMENTO',
  SEZIONE_22_PERFOMANCE = 'SEZIONE_22_PERFOMANCE',
  SEZIONE_23_DATIPUBBLICATI = 'SEZIONE_23_DATIPUBBLICATI',
  SEZIONE_23_ATTIVITASENSIBILE = 'SEZIONE_23_ATTIVITASENSIBILE',
  SEZIONE_23_EVENTORISCHIO = 'SEZIONE_23_EVENTORISCHIO',
  SEZIONE_4_SOTTOFASEMONITORAGGIO = 'SEZIONE_4_SOTTOFASEMONITORAGGIO',
  SEZIONE_4_CATEGORIAOBIETTIVI = 'SEZIONE_4_CATEGORIAOBIETTIVI',
  SEZIONE_APPROVAZIONE_PUBBLICAZIONE = 'SEZIONE_APPROVAZIONE_PUBBLICAZIONE',
  SEZIONE_INVIA_APPROVAZIONE = 'SEZIONE_INVIA_APPROVAZIONE',
  SEZIONE_MANCATA_COMPILAZIONE = 'SEZIONE_MANCATA_COMPILAZIONE',
  HELP_DESK = 'HELP_DESK',
  AVVISI_DFP = 'AVVISI_DFP',
}